Separation of models and adapter in independent packages. Changed share icon. Removed likes and like button. Changed comment to feedback.
