+++ app/src/main/java/com/example/pianoafrik/mestcollabsapp/Globall.java
@@ -3,12 +3,8 @@ package com.example.pianoafrik.mestcollabsapp;
 import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.os.Environment;
-import android.telephony.TelephonyManager;
-import android.util.Log;
 
-import org.json.JSONArray;
-import org.json.JSONObject;
+import com.example.pianoafrik.mestcollabsapp.models.Feeds;
 
 import java.util.ArrayList;
 import java.util.List;
